Move one to common/functions.py (#1007)
